backend/trynewbot/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI & ML
torch>=2.0.0
torchaudio>=2.0.0
transformers==4.35.0
openai-whisper==20231117
pyannote.audio==3.1.1
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3

# Zoom SDK (optional - requires separate setup)
# zoomus==1.1.5

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1